Refactor: main에서 사용되던 App, route, render를 분리하여 의존도를 낮추고, render만 선언합니다. TODO: main에서 이벤트위임을 하는 방식을 적용하는데 이벤트 위임 역시 분리가 가능한지 테스트
