+++ src/main.js
@@ -1,36 +1,14 @@
-import { MainPage, LoginPage, ErrorPage, ProfilePage } from "./page";
+import { render } from "./lib/render";
 import { clearUser, setUser } from "./store/user";
 
 const root = document.getElementById("root");
 
-const App = () => {
-  if (location.pathname === "/") return MainPage();
-  if (location.pathname === "/profile") return ProfilePage();
-  if (location.pathname === "/login") return LoginPage();
-  return ErrorPage();
-};
-
-const render = () => {
-  root.innerHTML = App();
-  route();
-};
-
-const route = () => {
-  root.querySelectorAll("a").forEach((el) => {
-    el.addEventListener("click", (e) => {
-      e.preventDefault();
-      const nextPathName = e.target.href.replace(location.origin, "");
-      history.pushState(null, "", nextPathName);
-      render();
-    });
-  });
-};
+render();
 
 window.addEventListener("popstate", () => {
   render();
 });
 
-//login
 root.addEventListener("submit", function (e) {
   e.preventDefault();
 
@@ -54,10 +32,8 @@ root.addEventListener("submit", function (e) {
 
 root.addEventListener("click", function (e) {
   if (e.target && e.target.id === "logout") {
-    console.log("로그아웃");
     clearUser();
     history.pushState(null, "", "login");
     render();
   }
 });
-render();
